Add files via upload
The "Image augmentation" python program is used to import the image data from your device and resize the image to an (80,80) pixel image and convert the image from color to grayscale. When using the program change the value of "root_dir" to reflect the directory of where your image data is located. Note that the program only works for one image folder at a time so your directory must be the folder that directly contains the image data. The print statement in cell 2 will print out the names of files contained in the "root_dir" variable. confirm with this output to check the names of files the program will be augmenting.