Add files via upload
The "KNN_code" python program is used to analyzed the augmented image data from the "t1" zip folder. the program converts the image data to a NumPy array and assigns the image a class. The sklearn python library splits the NumPy image array and the class data into training and test groups and using the training group predicts the classes of the test group. The accuracy of the prediction is outputted and visualized using a confusion matrix.